refactor(workspace): fold rust/tasks into services + retire _nexus_tasks.so
Phase 3 restructure plan #6, no longer deferred.  The standalone
`tasks` cdylib (`_nexus_tasks.so`) is folded into the services rlib so
the runtime ships a single Python wheel — `nexus_kernel` —
surfacing every Rust pyclass.  Tasks belong in services tier per the
parent plan's `services ⊥ backends ⊥ transport ⊥ raft` peer-crate
invariant; they were a separate crate purely for historical reasons.

## File moves (git mv)

  rust/tasks/src/engine.rs    → rust/services/src/tasks/engine.rs
  rust/tasks/src/error.rs     → rust/services/src/tasks/error.rs
  rust/tasks/src/priority.rs  → rust/services/src/tasks/priority.rs
  rust/tasks/src/retry.rs     → rust/services/src/tasks/retry.rs
  rust/tasks/src/store.rs     → rust/services/src/tasks/store.rs
  rust/tasks/src/task.rs      → rust/services/src/tasks/task.rs
  rust/tasks/src/lib.rs       → rust/services/src/tasks/mod.rs

In `mod.rs`: dropped the `#[global_allocator]` (rlibs can't declare
one — the cdylib owns mimalloc) and the `#[pymodule] fn _nexus_tasks`
envelope.  Pyclasses (PyTaskEngine / PyTaskRecord / PyQueueStats)
register through a new `pub fn register_python(m: &Bound<PyModule>)`.
`crate::*` imports inside the moved files become `super::*` to match
the new submodule nesting.

## Workspace + crate metadata

  Cargo.toml: drop `"rust/tasks"` workspace member.
  rust/services/Cargo.toml: add fjall + bincode + thiserror as
    runtime deps (formerly tasks's own); add tempfile as dev-dep
    so the moved unit tests compile.
  rust/services/src/lib.rs: add `pub mod tasks;` (gated behind the
    `python` feature, mirroring `permission`).
  rust/services/src/python/mod.rs: `register()` calls
    `crate::tasks::register_python(m)?` so the cdylib exposes the
    same pyclass surface.
  rust/tasks/: directory deleted (Cargo.lock + Cargo.toml + pyproject.toml).

## Python boot path

  src/nexus/tasks/__init__.py: imports of `_nexus_tasks` rewritten to
  `nexus_kernel`.  Module docstring updated to flag the consolidation.

## Verification

  cargo check --workspace --tests: clean.
  cargo clippy --workspace --all-targets -- -D warnings: clean.
  cargo fmt --all: applied.
