+++ rust/services/Cargo.toml
@@ -34,6 +34,10 @@ raft = { workspace = true }
 # syscalls (the same path Python takes), never via direct cross-crate
 # imports.  See `services/src/lib.rs` for the rationale.
 redb = "2"
+# Phase 3 restructure plan #6: tasks crate folded in here.
+fjall = "3"
+bincode = "1.3"
+thiserror = "2.0"
 dashmap = "6.1"
 parking_lot = "0.12"
 serde = { workspace = true }
@@ -43,3 +47,6 @@ tracing = "0.1"
 # pyo3 feature unioned by cargo when nexus-cdylib enables `extension-module`
 # on its own pyo3 dep — services does not need to set extension-module here.
 pyo3 = { version = "0.27.2", features = ["hashbrown"], optional = true }
+
+[dev-dependencies]
+tempfile = "3"
