--- rust/tasks/pyproject.toml
@@ -1,17 +1,0 @@
-[build-system]
-requires = ["maturin>=1.0,<2.0"]
-build-backend = "maturin"
-
-[project]
-name = "nexus-tasks"
-version = "0.1.0"
-description = "Durable task queue engine for Nexus Agent OS"
-requires-python = ">=3.14"
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-]
-
-[tool.maturin]
-module-name = "_nexus_tasks"
-bindings = "pyo3"
